Print line number and function name in addition to file name.
This commit uses Chrome's 'error.stack' thing to print more information about the console.log call, mostly WHERE IT WAS BLOODY WELL FROM. (That is to say, "WorkerConsole.js:54" is not a very informative line number for my error.) This is ridiculously chrome-specific, though, so it breaks compatibility for world+dog. The new format is 'function name' + 'shorter url' + 'line number'. This works even if you reassigned "console" to something shorter, like "c".